Add import tests to the workers project
These tests make sure that we can import all of the test files by
dynamically importing each one when the tests are called with the
appropriate tag. This fixes a problem we had where the compendia tests
were silently not being run because of an issue with one of our
dependencies.