Add spearman correlations to the no_op tests
These work similarly to the ones in the affymetrix tests. Basically, we
download expected data from S3 and then run a spearman correlation
against the data that we processed.

Some of these have a slight problem that I will have to talk to the data
science team about though -- the spearman correlation fails because the
new and old versions of the processed file have different numbers of
rows. I think this has something to do with the genes, and maybe newer
versions picking out more genes? I will have to ask how we want to
handle this.