Start refactoring the spearman tests
This abstracts the spearman tests out into a separate
`assertMostlyAgrees` helper function, so that we can apply this to every
test.

I also swapped out one of the no_header files, because the original one
used looks to have been contrived because the actual source file is an
affy file that needs to be processed. This meant that the file I
downloaded from refine.bio to compare against didn't work so well.

Now we are successfully able to compare data frames of different sizes
against each other, but now we have a new problem: the join that we are
doing is messing up the spearman correlation because for some reason
some genes have multiple observations:

$ cat gene_converted_GSM557500-tbl-1.txt | sort | tail -n 3

    ENSG00000284741 6.355069
    ENSG00000284741 6.3770933
    ENSG00000284741 6.442038

I will have to figure out how to handle this early next week.