Add a quality check to the end-to-end tests
Unfortunately, this check fails for now, with a correlation of only
~57%. I tried this with both the .sra file and the .fastq files and both
of them gave approximately the same correlation, so there must be some
other problem.

It is suspicious to me that so few of the genes line up with the version
I grabbed from the compendium; I wonder if I will pick up more genes if
I gene convert them?

There could be some other issue wrong with the samples; I am not
entirely sure. I am also not 100% sure that my quant file is correct, I
just pulled it from the compendium from refine.bio.