Fix a bug in the salmon processor
When writing the end-to-end tests, I kept running into an issue where
the tests would fail with the cryptic error message of "file already
exists". After trying to debug this for a while, I finally figured out
that the issue was that we had a broken link. Because the link is
broken, os.path.exists() returns False, but because the link exists any
attempts to create a new symbolic link to that path name would fail.
This solves that problem.