Try removing the version in the transcript ids
Instead of mapping everything to gene IDs, which is a little bit iffy
because we are dealing with potentially different transcripts for
different mutations of the same gene, let's try sticking with transcript
IDs instead and just stripping the version suffix.

It turns out that this approach doesn't really work, because we get data
that looks something like this:
> B0024.9a.1	189	0
> B0024.9a.2	0	189
which the gene ID filter fixes but this does not.

When I make the PR we will have to have a discussion on which approach
we should take to comparing quant.sf files.