Fix the checks in the compendia test
Because I fixed the typo `RNASEQ` -> `RNA-SEQ` the checks I implemented
stopped working. After tracing the issue back to the above typo change,
I figured out that the problem was that we expected some samples to be
filtered out, so I checked to make sure the expected behavior actually
happened.